# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Wypożyczalnia Aut - API
  description: API do obsługi wypożyczalni
  version: "1.0"
servers:
  - url: http://localhost:5000/
...
paths:
  /login/login:
    post:
      tags:
        - login
      summary: Loguje do systemu
      description: Loguje uzytkownka do systemu
      operationId: login_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login_login_body'
      responses:
        "200":
          description: Token sesji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: Niewlasciwe haslo lub login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
...
  /login/logout:
    post:
      tags:
        - login
      summary: Wyloguj z systemu
      operationId: login_logout_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      responses:
        "200":
          description: Wylogowano
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.login_controller
...
components:
  schemas:
    Empty:
      type: object
      properties: { }
      description: Pusty Json
      example: { }
...
    login_login_body:
      type: object
      properties:
        login:
          type: string
          description: Login
          example: TEstasdasd
        password:
          type: string
          description: "Hasło, zahashowane bCryptem"
    inline_response_200:
      type: object
      properties:
        token:
          $ref: '#/components/parameters/SessionToken'
      example:
        token: "{}"
...
  parameters:
    SessionToken:
      name: Session_Token
      in: header
      description: Token sesji
      required: true
      style: simple
      explode: false
      schema:
        type: integer
        example: 12943854693802
      example: "1111111111111111"
...